Move OGC and FIS enums to the api files
